Rename generator middleware file
The name implied it only served async middleware. Times have changed,
synchronous iterators are cool too.
