+++ src/index.ts
@@ -1,5 +1,5 @@
 export { default as createAction } from './actions/create-action';
 export { default as createReducer } from './reducers/create-reducer';
-export { default as middleware } from './middleware/async-generator-middleware';
+export { default as middleware } from './middleware/generator-middleware';
 export { mixin as swallow } from './utils/errors';
 export { nothing } from 'immer';
