+++ src/middleware/generator-middleware.ts
